tests and state changes

+++ programs/hxui/src/states/free_tokens_counter.rs
@@ -4,6 +4,6 @@ use anchor_lang::prelude::*;
 #[derive(InitSpace)]
 pub struct FreeTokensCounter{
     pub current_epoch:u64,
-    pub remaining_free_tokens:u8,
+    pub remaining_free_tokens:u64,
     pub bump:u8
 }
